+++ test/time_zone_info/data_store/erlang_term_storage_test.exs
@@ -56,10 +56,10 @@ defmodule TimeZoneInfo.DataStore.ErlangTermStorageTest do
 
     assert version == "2019c"
     assert time_zones == 387
-    assert links == 86
+    assert links == 205
     assert get_in(tables, [:time_zone_info, :size]) == 1
     assert get_in(tables, [:time_zone_info, :memory]) > 300
-    assert get_in(tables, [:time_zone_info_links, :size]) == 86
+    assert get_in(tables, [:time_zone_info_links, :size]) == 205
     assert get_in(tables, [:time_zone_info_links, :memory]) > 1600
     assert get_in(tables, [:time_zone_info_rules, :size]) == 29
     assert get_in(tables, [:time_zone_info_rules, :memory]) > 2100